fix(datalad): Improve snapshot exception handling
